Refactor generation of punctuator_table.h
Rename the script that generates punctuator_table.h and tweak it
slightly to be more consistent with the WIP scripts from the (hopefully)
upcoming Great Options Refactoring. Mostly, this uses file.write instead
of print, and argparse to grab the arguments. It also changes the
argument order and adds a CMake helper function to generate sources.
